refactor(#316): consolidate VLM parsing utilities, response models, and error classes
